im over this theres a bug with fave modal on other pages i dont feel like fixing
