+++ PlayFlix-firebase-FrontEnd-main/src/ApiKeys.js
@@ -9,3 +9,9 @@ export const firebaseConfig = {
   appId: "1:768528677602:web:d307610a8efad4e4bf3cbb"
 };
 
+export const searchGames = async (searchParams) => {
+  const search = searchParams.replace(/\s/g, '')   
+  const req = await fetch(`https://localhost:7215/api/Games/search?query=${search}`);
+  const resp = await req.json();
+  return await resp
+}
