+++ PlayFlix-firebase-FrontEnd-main/src/components/games/Game.js
@@ -7,10 +7,10 @@ import { Button, Modal } from "@mui/material/";
 export const Game = ({ game, onGameClick }) => {
   return (
     <>
-      <div>{game.title}</div>
-      <img src={game.gameImg}></img>
+      <div>{game?.title}</div>
+      <img src={game?.gameImg}></img>
       <Button
-        onClick={() => onGameClick(game.id)}
+        onClick={() => onGameClick(game?.id)}
         variant="contained"
         color="primary"
       >
