+++ PlayFlix-firebase-FrontEnd-main/src/components/views/ApplicationViews.js
@@ -6,12 +6,23 @@ import { Favorites } from "../favorites/Favorites";
 import { GamePlay } from "../games/GamesPlay";
 import { GamesLeaderBoard } from "../games/GamesLeaderBoard/Leaderboard";
 import { useState } from "react";
+import { searchGames } from "../../ApiKeys";
 
 export const ApplicationViews = ({navigate, setUserCheck, userCheck}) => {
   const[searchParams, setSearchParams] = useState("")
   const[modalOpen, setModalOpen] = useState(false)
   const [returnedGames, setReturnedGames] = useState([])
   
+  const handleClick = async () => {
+    try {
+      const searchedGames = await searchGames(searchParams)
+      setReturnedGames(searchedGames)
+      setModalOpen(true)
+      
+      } catch (error) {
+      console.log(error)
+      }
+    }
 
   return (
     
@@ -21,7 +32,9 @@ export const ApplicationViews = ({navigate, setUserCheck, userCheck}) => {
         path="/"
         element={
           <>
-          <NavBar navigate={navigate}
+            <NavBar
+            handleClick={handleClick}
+            navigate={navigate}
             userCheck={userCheck} 
             setUserCheck={setUserCheck} 
             setSearchParams={setSearchParams}
